+++ examples/H2_matrix_2level_construct.cpp
@@ -10,3 +10,30 @@
 #include "Hatrix/Hatrix.h"
 
 using randvec_t = std::vector<std::vector<double> >;
+
+namespace Hatrix {
+  class H2 {
+  private:
+    int64_t N, rank, admis;
+  public:
+    H2(const randvec_t& randvec, int64_t N, int64_t rank, int64_t admis) :
+      N(N), rank(rank), admis(admis) {
+    }
+  };
+}
+
+int main(int argc, char* argv[]) {
+  int64_t N = atoi(argv[1]);
+  int64_t rank = atoi(argv[2]);
+  int64_t admis = atoi(argv[3]);
+
+  Hatrix::Context::init();
+  randvec_t randpts;
+  randpts.push_back(Hatrix::equally_spaced_vector(N, 0.0, 1.0 * N)); // 1D
+  randpts.push_back(Hatrix::equally_spaced_vector(N, 0.0, 1.0 * N)); // 2D
+  randpts.push_back(Hatrix::equally_spaced_vector(N, 0.0, 1.0 * N)); // 3D
+
+  Hatrix::H2 A(randpts, N, rank, admis);
+
+  Hatrix::Context::finalize();
+}
