+++ include/Hatrix/util/helpers.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <vector>
+
+namespace Hatrix {
+  std::vector<double> equally_spaced_vector(int N, double minVal, double maxVal);
+}
