+++ src/util/Makefile
@@ -1,5 +1,5 @@
 TOPSRCDIR = ../..
 sublib  := libutil.a
-SOURCES := context.cpp matrix_generators.cpp
+SOURCES := context.cpp matrix_generators.cpp helpers.cpp
 
 include $(TOPSRCDIR)/common.mk
